% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enderstofighi.R
\name{compute_enderstofighi}
\alias{compute_enderstofighi}
\title{Center data using Enders & Tofighi (2007) method}
\usage{
compute_enderstofighi(data, sub, outcome, expect, ses, run)
}
\arguments{
\item{data}{A dataframe containing the input data.}

\item{sub}{The name of the column in 'data' representing subject IDs.}

\item{outcome}{The name of the column in 'data' representing outcome measures.}

\item{expect}{The name of the column in 'data' representing expected values.}

\item{ses}{The name of the column in 'data' representing session IDs.}

\item{run}{The name of the column in 'data' representing run numbers.}
}
\value{
A dataframe processed with added statistical measures.
  - `OUTCOME`: Numeric version of the outcome column.
  - `EXPECT`: Numeric version of the expect column.
  - `OUTCOME_cm`: Mean of `OUTCOME` for each subject.
  - `OUTCOME_demean`: `OUTCOME` demeaned for each data point.
  - `EXPECT_cm`: Mean of `EXPECT` for each subject.
  - `EXPECT_demean`: `EXPECT` demeaned for each data point.
  - `trial_index`: Sequential trial number for each subject.
  - `lag.OUTCOME_demean`: Lagged version of `OUTCOME_demean`.
  - `EXPECT_cmc`: Centered mean contrast for `EXPECT`.
  - The function returns a subset of data with complete cases for `lag.OUTCOME_demean`.
}
\description{
This function processes the input data by calculating several statistical measures
including outcome averages, demeaned outcomes, and trial indices. It also computes
lagged outcomes and contrast measures for behavioral analysis.
}
\examples{
# Assuming data is your dataframe with appropriate columns:
# processed_data <- process_main_data(data, "src_subject_id",
#                                    "event04_actual_angle", "event02_expect_angle",
#                                    "session_id", "param_run_num")

}
\references{
Main text: Enders & Tofighi (2007). Centering predictor variables in cross-sectional multilevel models: 
  A new look at an old issue. Psychological Methods, 12(2), 121–138.
  https://doi.org/10.1037/1082-989X.12.2.121
Masur, P. (2018). How to Center in Multilevel Models. 
  Retrieved from https://philippmasur.de/2018/05/23/how-to-center-in-multilevel-models/
Shaw. M., & Flake, J. K. . Module 8: Centering in MLMs. 
  Retrieved from https://www.learn-mlms.com/08-module-8.html
}
