% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ICC.R
\name{compute_icc}
\alias{compute_icc}
\title{Compute Intraclass Correlation Coefficient (ICC) from a Linear Mixed Model}
\usage{
compute_icc(lmer_object)
}
\arguments{
\item{lmer_object}{A linear mixed-effects model object, typically created 
using `lmer()` function from the `lme4` package.}
}
\value{
A numeric value representing the ICC of the fitted model.
}
\description{
This function calculates the Intraclass Correlation Coefficient (ICC) from 
a fitted linear mixed-effects model object (from the `lme4` package). The ICC 
is a measure of the proportion of the total variance that is attributable to 
the grouping structure in the data.
}
\examples{
# Example usage:
# model <- lmer(response ~ predictor + (1|group), data = dataset)
# icc_value <- compute_icc(model)

}
