% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_randomeffects.R
\name{compute_randomeffects}
\alias{compute_randomeffects}
\title{Compute Random and Fixed Effects from a Linear Mixed Effects Model}
\usage{
compute_randomeffects(
  lmer_model,
  rand_savefname,
  taskname,
  new_rand_names = c("rand_intercept", "rand_cue"),
  new_fix_names = c("fix_intercept", "fix_cue")
)
}
\arguments{
\item{lmer_model}{A linear mixed effects model object from which to extract effects.}

\item{rand_savefname}{The file path where the combined random and fixed effects data frame should be saved.
If NULL, the file is not saved.}

\item{taskname}{The name of the task associated with the model, to be added as a new column in the output.}

\item{new_rand_names}{A character vector specifying new names for the random effects terms.}

\item{new_fix_names}{A character vector specifying new names for the fixed effects terms.}
}
\value{
An invisible data frame combining the processed random and fixed effects, along with the task name.
}
\description{
This function takes a linear mixed effects model object, extracts the fixed effects and random effects,
processes them, combines them into a single data frame, and optionally saves the result to a file.
}
\examples{
# Assuming `model` is a lmer model object and `task_name` is the associated task
result_df <- compute_randomeffects(model, "path/to/savefile.csv", task_name)

# If you do not wish to save the result to a file, use NULL for `rand_savefname`
result_df <- compute_randomeffects(model, NULL, task_name)

}
