% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summary_NPS.R
\name{compute_summary_NPS}
\alias{compute_summary_NPS}
\title{Summarize NPS extraction data for Plotting}
\usage{
compute_summary_NPS(
  df,
  groupwise_measurevar,
  subject_keyword,
  model_iv1,
  model_iv2
)
}
\arguments{
\item{df}{A data frame containing the neurophysiological data.}

\item{groupwise_measurevar}{The name of the column in `df` that contains the variable for 
which group-wise summary statistics are to be calculated.}

\item{subject_keyword}{The name of the column in `df` that identifies individual subjects.}

\item{model_iv1}{The first independent variable for the model, used for grouping in the analysis.}

\item{model_iv2}{The second independent variable for the model, used for grouping in the analysis.}
}
\value{
A list containing two elements: `subjectwise` and `groupwise`. 
  `subjectwise` is a data frame with subject-level summary statistics, 
  and `groupwise` is a data frame with group-level summary statistics.
}
\description{
This function preprocesses and summarizes neurophysiological data for plotting purposes. 
It categorizes and orders cue and stimulus levels, then calculates subject-wise and group-wise
summary statistics for a given measure.
}
\examples{
# Example usage:
# NPS_summary_for_plots(data_frame, "measure_var", "subject_id", "iv1", "iv2")

}
