% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_filter_ses_trial.R
\name{df_filter_ses_trial}
\alias{df_filter_ses_trial}
\title{Filter dataframe using session and trial number as criterion}
\usage{
df_filter_ses_trial(
  data,
  session_colname,
  subject_colname,
  session_threshold,
  trial_threshold
)
}
\arguments{
\item{data}{dataframe}

\item{session_colname}{within dataframe, column name that contains information on session id}

\item{subject_colname}{within dataframe, column name that contains information on subject id}

\item{session_threshold}{int: What is the cutoff of the dataframe? subjects with all three sessions?}

\item{trial_threshold}{int: bare minimum number of trials that each participant should have.}
}
\value{
dataframe with filtered participants that match trial_threshold, session_threshold criterion
}
\description{
Filter dataframe using session and trial number as criterion
}
