% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_load_beh.R
\name{df_load_beh}
\alias{df_load_beh}
\title{Load and Process Behavioral Data for Specified Task}
\usage{
df_load_beh(datadir, taskname, subject_varkey, iv, dv, exclude)
}
\arguments{
\item{datadir}{A string specifying the directory path where the data is stored.}

\item{taskname}{A string specifying the name of the task (e.g., 'pain', 'cognitive', 'vicarious').}

\item{subject_varkey}{A string specifying the variable key for subjects 
(e.g., 'src_subject_id' or 'subject').}

\item{iv}{A string specifying the independent variable in the analysis.}

\item{dv}{A string specifying the dependent variable in the analysis. 
If provided, the function will return data with NA values in this column removed.}

\item{exclude}{A vector specifying any criteria for excluding data.}
}
\value{
A dataframe containing the loaded and processed data. If 'dv' is specified,
  returns data with NA values in the 'dv' column removed.
}
\description{
This function loads behavioral data for a specified task from a given directory,
binds them into a single dataframe, and processes the data according to specified criteria.
It handles missing values, factors, and optionally filters the data based on a dependent variable.
this code allows the function to behave differently based on whether a specific column name (dv) is provided or not.
}
\details{
Regarding dv,
If dv is given, the function returns the original dataframe but with rows containing missing values in the dv column removed. 
If dv is not provided, the function returns the original dataframe as is. 
This functionality is useful when you want to make the data cleaning step conditional based on the presence of a specific variable.
}
\examples{
# Assuming appropriate data and parameters
# df <- df_load_beh(datadir = "path/to/data", taskname = "pain",
#                  subject_varkey = "subjectID", iv = "independentVar", 
#                  dv = "dependentVar", exclude = c("criteria1", "criteria2"))
}
