% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_load_signature.R
\name{df_load_signature}
\alias{df_load_signature}
\title{Load data frame that has signature extractions.}
\usage{
df_load_signature(fname)
}
\arguments{
\item{fname}{str: filename glob pattern that comes after the signature_key (e.g. "_sub-all_runtype-pvc_event-stimulus.tsv")}
}
\value{
dataframe with signature-extracted dot products, filenames are split into informative columns
}
\description{
Signature extractions are computed via the following code. 
Here, we load the derivatives.
https://github.com/jungheejung/cue_expectancy/tree/main/scripts/step08_applyNPS
https://github.com/jungheejung/cue_expectancy/blob/main/scripts/step10_nilearn/singletrialLSS/step02_apply_signature.py
}
\details{
load .tsv or .csv file that list signature extracted dot products
Note that the filenames have the metadata, thereby, we extract the metadata and include it as new columns.
}
\examples{
df_load_signature(fname)
}
