% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_rename_columns.R
\name{df_rename_columns}
\alias{df_rename_columns}
\title{Rename DataFrame Columns Based on Mapping}
\usage{
df_rename_columns(data, column_mapping)
}
\arguments{
\item{data}{A DataFrame.}

\item{column_mapping}{A named character vector where names represent old column names and values represent new column names.}
}
\value{
A modified DataFrame with new columns having the same old names but renamed values if applicable.
}
\description{
This function creates new columns in a DataFrame with the same old names but renamed values based on a mapping of old names to new names.
}
\examples{
df <- data.frame(
  event04_actual_angle = c(45, 60, 30),
  event02_expect_angle = c(1, 2, 3),
  other_column = c("A", "B", "C")
)
column_mapping <- c("event04_actual_angle" = "OUTCOME", "event02_expect_angle" = "EXPECT")
df_with_renamed_columns <- df_rename_columns(df, column_mapping)

}
