% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_hline_bartoshuk.R
\name{ggplot_hline_bartoshuk}
\alias{ggplot_hline_bartoshuk}
\title{Add horizontal lines and labels to a plot.}
\usage{
ggplot_hline_bartoshuk(
  g,
  xposition,
  linesize = 0.1,
  linetype = "dashed",
  linecolor = "#adb5bd",
  nudge_x = 0.1,
  textsize = 9
)
}
\arguments{
\item{xposition}{The x-position where the lines and labels will be added.}
}
\value{
NULL. This function is used for side effects (adding lines and labels to a plot).
}
\description{
This function adds horizontal lines and labels to a plot at predefined
y-intercepts with corresponding labels.
}
\examples{
p <- ggplot(data, aes(x = x, y = y)) + geom_point()
p + ggplot_hline_bartoshuk(3.5)

}
