% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_prettifylegend.R
\name{ggplot_prettifylegend}
\alias{ggplot_prettifylegend}
\title{Prettify a Legend and Combine with a Plot}
\usage{
ggplot_prettifylegend(
  ggplot_obj,
  factor_level,
  factor_color,
  geom_point_size = 3,
  legend_position = c(-0.1, 0.9),
  legend_widths = c(4, 1)
)
}
\arguments{
\item{ggplot_obj}{A ggplot object that you want to combine with the legend.}

\item{factor_level}{A character vector specifying the factor levels for the legend.}

\item{factor_color}{A character vector specifying the colors for each factor level.}

\item{geom_point_size}{Size of the points in the legend. Default is 3.}

\item{legend_position}{A numeric vector specifying the legend's x and y positions. Default is c(-0.1, 0.9).}

\item{legend_widths}{A numeric vector specifying the relative widths of the plot and legend. Default is c(4, 1).}
}
\value{
A combined plot with a customized legend.
}
\description{
This function takes a ggplot object, customizes a legend, and combines it with the plot.
}
\examples{
\dontrun{
my_ggplot <- prettify_legend(my_ggplot, factor_level = c("High cue", "Low cue"), factor_color = c("#D73027", "#4575B4"))
}

}
