% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_extraction.R
\name{load_extraction}
\alias{load_extraction}
\title{load_signature_extraction}
\usage{
load_extraction(fname)
}
\arguments{
\item{fname}{str: filename glob pattern that comes after the signature_key (e.g. "_sub-all_runtype-pvc_event-stimulus.tsv")}
}
\value{
dataframe with signature-extracted dot products, filenames are split into informative columns
}
\description{
load .tsv or .csv file that list signature extracted dot products
}
\examples{
load_signature_extraction(fname)
}
