% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSummary.R
\name{meanSummary}
\alias{meanSummary}
\title{Calculate Mean and Standard Deviation for Groups}
\usage{
meanSummary(data, group, dv)
}
\arguments{
\item{data}{A data frame containing the data to be summarized.}

\item{group}{The grouping variable as a string or a vector of strings if multiple grouping variables are used.}

\item{dv}{The name of the column in `data` that contains the dependent variable for which the mean and standard deviation are calculated.}
}
\value{
A data frame with each row representing a group defined by `group`. Each row contains
  the mean (`mean_per_sub`) and standard deviation (`sd`) of the specified dependent variable `dv` for that group.
}
\description{
This function computes the mean and standard deviation for a specified dependent variable (dv),
grouped by a specified variable in a data frame. It returns a data frame with these summary statistics.
}
\examples{
# Example usage with the built-in mtcars dataset, grouping by 'cyl' and summarizing 'mpg'
meanSummary(mtcars, "cyl", "mpg")

}
