% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSummary_2continuous.R, R/meanSummary_2dv.R
\name{meanSummary_2continuous}
\alias{meanSummary_2continuous}
\title{Calculate Summary Statistics for Two Continuous Variables}
\usage{
meanSummary_2continuous(DATA, GROUP, DV1, DV2)

meanSummary_2continuous(DATA, GROUP, DV1, DV2)
}
\arguments{
\item{DATA}{A dataframe containing the data for analysis.}

\item{GROUP}{A vector of column names or a single column name used for grouping the data.}

\item{DV1}{The name of the first continuous variable (dependent variable) for which summary statistics are calculated.}

\item{DV2}{The name of the second continuous variable (dependent variable) for which summary statistics are calculated.}
}
\value{
A dataframe where each row represents a group defined in `GROUP` and contains the mean and standard deviation for `DV1` and `DV2`.

A dataframe where each row represents a group defined in `GROUP` and contains the mean and standard deviation for `DV1` and `DV2`.
}
\description{
This function computes summary statistics for two continuous variables within specified groups in a dataset. It calculates the mean and standard deviation for each variable in each group.

This function computes summary statistics for two continuous variables within specified groups in a dataset. It calculates the mean and standard deviation for each variable in each group.
}
\examples{
# Assuming DATA is your dataframe with appropriate columns and groups:
# summary_stats <- meanSummary_2continuous(DATA, c("Group1", "Group2"),
#                                         "Variable1", "Variable2")

# Assuming DATA is your dataframe with appropriate columns and groups:
# summary_stats <- meanSummary_2continuous(DATA, c("Group1", "Group2"),
#                                         "Variable1", "Variable2")

}
