% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_continuousbin.R
\name{plot_binned_rating}
\alias{plot_binned_rating}
\title{Plot Binned Ratings}
\usage{
plot_binned_rating(
  df,
  taskname,
  iv1,
  iv2,
  subject = "src_subject_id",
  xlab = "expectation ratings",
  ylab = "outcome ratings",
  levels = 10
)
}
\arguments{
\item{df}{A dataframe containing the data to be plotted.}

\item{taskname}{A character string specifying the task name, used for filtering data and generating plot titles.}

\item{iv1}{The name of the first independent variable column in `df`.}

\item{iv2}{The name of the second independent variable column in `df`.}

\item{subject}{The name of the subject identifier column in `df` (default is "src_subject_id").}

\item{xlab}{The label for the x-axis (default is "expectation ratings").}

\item{ylab}{The label for the y-axis (default is "outcome ratings").}

\item{levels}{The number of levels for binning the independent variable (default is 10).}
}
\value{
A ggplot object representing the binned rating plot.
}
\description{
This function generates plots for binned ratings in a specified task, with options for customizing
color gradients and labels. It supports different task types with distinct color palettes. The function
assumes specific column names in the data and uses custom ggplot formatting functions.
}
\examples{
# Example usage
plot_binned_rating(df, "pain", "event02_expect_angle", "event04_actual_angle",
                   "src_subject_id", "expectation ratings", "outcome ratings", 10)
# Note: `df` should be a properly formatted dataframe as required by the function.
#### [Example](https://github.com/jungheejung/cueR/blob/main/man/figures/example_plot_binned_rating.png)
}
