% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cueexpectancy_onefactor.R
\name{plot_cueexpectancy_onefactor}
\alias{plot_cueexpectancy_onefactor}
\title{Plot Cue Expectancy with One Factor}
\usage{
plot_cueexpectancy_onefactor(
  subjectwise,
  groupwise,
  iv,
  sub_mean,
  group_mean,
  se,
  subject,
  ggtitle,
  title,
  xlab,
  ylab,
  taskname,
  w,
  h,
  dv_keyword,
  color,
  plot_savefname,
  expand_x = TRUE,
  xlim = c(NA, 3),
  ylim = c(NA, NA),
  x_scale_expansion = c(0, 0.5),
  x_hline_position = 3.5,
  x_hline_linetype = "dashed",
  x_hline_nudge_x = 0.1,
  x_hline_textsize = 3,
  legend_factor_levels = c("High cue", "Low cue"),
  legend_factor_colors = c("#D73027", "#4575B4"),
  legend_geom_point_size = 3,
  legend_position = c(-0.1, 0.9),
  legend_widths = c(4, 1)
)
}
\arguments{
\item{subjectwise}{A data frame containing subject-wise data.}

\item{groupwise}{A data frame containing group-wise summary data.}

\item{iv}{Independent variable name (as a string) in the data frame.}

\item{sub_mean}{Name of the subject-wise mean variable in the data frame.}

\item{group_mean}{Name of the group mean variable in the data frame.}

\item{se}{Standard error for the group mean.}

\item{subject}{Name of the subject identifier variable in the data frame.}

\item{ggtitle}{Title for the ggplot.}

\item{title}{Title for the legend.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{taskname}{Name of the task (unused in function).}

\item{w}{Width of the saved plot.}

\item{h}{Height of the saved plot.}

\item{dv_keyword}{Keyword associated with the dependent variable (unused in function).}

\item{color}{Color scheme for the plot, specified as a vector of colors.}

\item{plot_savefname}{Filename for saving the plot.}

\item{expand_x}{Logical, whether to expand x-axis.}

\item{xlim}{Limits for the x-axis (vector of two values).}

\item{ylim}{y-axis limits for the plot.}

\item{x_scale_expansion}{Expansion factor for x-axis scaling.}

\item{x_hline_position}{Position for horizontal line.}

\item{x_hline_linetype}{Linetype for the horizontal line.}

\item{x_hline_nudge_x}{Nudge x value for the horizontal line.}

\item{x_hline_textsize}{Text size for labels near the horizontal line.}

\item{legend_factor_levels}{Factor levels for the legend.}

\item{legend_factor_colors}{Colors for the legend factors.}

\item{legend_geom_point_size}{Size of the points in the legend.}

\item{legend_position}{Position of the legend (vector of two values).}

\item{legend_widths}{Widths of the legend components (vector of two values).}
}
\value{
A ggplot object representing the customized cue expectancy plot.
}
\description{
This function creates a customized plot for cue expectancy data with one factor. It builds upon the `plot_halfrainclouds_onefactor` function and adds additional customizations like horizontal lines, legend modifications, and axis adjustments.
}
\examples{
# Example usage (assuming appropriate data frames 'subjectwise' and 'groupwise'):
plot_cueexpectancy_onefactor(subjectwise, groupwise, "IVName",
                            "SubMean", "GroupMean", "SE", "SubjectID",
                            "Plot Title", "Legend Title", "X-axis Label", "Y-axis Label",
                            "TaskName", c(-2, 2), 8, 6, "DVKeyword",
                            c("blue", "red"), "output_plot.png",
                            TRUE, c(NA, 3), c(0, 0.5), 3.5, "dashed", 0.1, 3,
                            c("High cue", "Low cue"), c("#D73027", "#4575B4"), 3,
                            c(-0.1, 0.9), c(4, 1))

![Example Plot](https://github.com/jungheejung/cueR/blob/main/man/figures/example_plot_cueexpectancy_twofactor.png)
}
