% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cueexpectancy_twofactor.R
\name{plot_cueexpectancy_twofactor}
\alias{plot_cueexpectancy_twofactor}
\title{Create a Cue Expectancy Plot with Two Factors}
\usage{
plot_cueexpectancy_twofactor(
  subjectwise,
  groupwise,
  model_iv1,
  model_iv2,
  sub_mean,
  group_mean,
  se,
  subject,
  ggtitle,
  title,
  xlab,
  ylab,
  taskname,
  ylim,
  w,
  h,
  dv_keyword,
  color,
  plot_savefname,
  expand_x = TRUE,
  xlim = c(NA, 3),
  x_scale_expansion = c(0, 0.5),
  x_hline_position = 3.5,
  x_hline_linetype = "dashed",
  x_hline_nudge_x = 0.1,
  x_hline_textsize = 3,
  legend_factor_levels = c("High cue", "Low cue"),
  legend_factor_colors = c("#D73027", "#4575B4"),
  legend_geom_point_size = 3,
  legend_position = c(-0.1, 0.9),
  legend_widths = c(4, 1)
)
}
\arguments{
\item{subjectwise}{Data for outcome by subject.}

\item{groupwise}{Data for outcome by group.}

\item{model_iv1}{First independent variable for modeling.}

\item{model_iv2}{Second independent variable for modeling.}

\item{sub_mean}{Mean outcome by subject.}

\item{group_mean}{Mean outcome by group.}

\item{se}{Standard error.}

\item{subject}{Subject information.}

\item{ggtitle}{Title for the plot.}

\item{title}{Title for the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{taskname}{Task name.}

\item{ylim}{Limits for the y-axis.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{dv_keyword}{Keyword for dependent variable.}

\item{color}{Color for plot elements.}

\item{plot_savefname}{Filename for saving the plot.}

\item{expand_x}{Expand limits for the x-axis. Default is TRUE.}

\item{xlim}{Limits for the x-axis. Default is c(NA, 3).}

\item{x_scale_expansion}{Expansion for x-axis scale. Default is c(0, 0.5).}

\item{x_hline_position}{Position for horizontal line on the x-axis. Default is 3.5.}

\item{x_hline_linetype}{Linetype for the horizontal line. Default is "dashed".}

\item{x_hline_nudge_x}{Nudge value for horizontal line on the x-axis. Default is 0.1.}

\item{x_hline_textsize}{Text size for the horizontal line annotation. Default is 3.}

\item{legend_factor_levels}{Levels for the legend factors. Default is c("High cue", "Low cue").}

\item{legend_factor_colors}{Colors for the legend factors. Default is c("#D73027", "#4575B4").}

\item{legend_geom_point_size}{Size of the points in the legend. Default is 3.}

\item{legend_position}{Position of the legend. Default is c(-0.1, 0.9).}

\item{legend_widths}{Relative widths of the plot and legend. Default is c(4, 1).}
}
\value{
A customized cue expectancy plot with two factors.
}
\description{
This function generates a cue expectancy plot with two factors, including customization options for various plot elements.
}
\examples{
# Example usage:
    g <- plot_cueexpectancy_twofactor(
        outcome_subjectwise, outcome_groupwise, model_iv1, model_iv2,
        sub_mean, group_mean, se, subject, ggtitle, title,
        xlab, ylab, taskname, ylim,
        w, h, dv_keyword, color, plot_savefname,
        x_hline_position = 3.5,
        x_hline_linetype = "dashed",
        x_hline_nudge_x = 0.1,
        x_hline_textsize = 3,
        legend_factor_levels = c("High cue", "Low cue"),
        legend_factor_colors = c("#D73027", "#4575B4"),
        legend_geom_point_size = 4,
        legend_position = c(-0.1, 0.9),
        legend_widths = c(4, 1)
        )
        print(g)
        ggsave(plot_savefname, width = w, height = h)

![Example Plot](https://github.com/jungheejung/cueR/blob/main/man/figures/example_plot_cueexpectancy_twofactor.png)
}
