% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errorbar.R
\name{plot_errorbar}
\alias{plot_errorbar}
\title{Plot Error Bars}
\usage{
plot_errorbar(
  subjectwise,
  groupwise,
  iv,
  sub_iv,
  group_by,
  subjectwise_mean,
  group_mean,
  se,
  subject,
  ggtitle,
  title,
  xlab,
  ylab,
  taskname,
  ylim,
  w,
  h,
  dv_keyword,
  color,
  level_num,
  save_fname
)
}
\arguments{
\item{subjectwise}{Data frame containing subject-wise data.}

\item{groupwise}{Data frame containing group-wise summary data.}

\item{iv}{Name of the independent variable in the data frames.}

\item{sub_iv}{Sub independent variable in the data frames (unused in current implementation).}

\item{group_by}{Grouping variable name used in the data frames.}

\item{subjectwise_mean}{Name of the subject-wise mean variable in `subjectwise`.}

\item{group_mean}{Name of the group mean variable in `groupwise`.}

\item{se}{Name of the standard error variable in `groupwise`.}

\item{subject}{Subject identifier variable in the data frames (unused in current implementation).}

\item{ggtitle}{Title for the ggplot.}

\item{title}{Title for the legend.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{taskname}{Task name (unused in current implementation).}

\item{ylim}{Vector of two numbers for y-axis limits.}

\item{w}{Width of the saved plot.}

\item{h}{Height of the saved plot.}

\item{dv_keyword}{Dependent variable keyword (unused in current implementation).}

\item{color}{Vector of colors for the plot.}

\item{level_num}{Number of levels for the independent variable.}

\item{save_fname}{File name to save the plot.}
}
\value{
A ggplot object representing the error bars plot.
}
\description{
Creates a plot with error bars and points, based on subject-wise and group-wise data.
The plot is customized with various aesthetic parameters and saved to a file.
}
\examples{
# Assuming subjectwise and groupwise are appropriately structured data frames:
plot_errorbar(subjectwise, groupwise, "iv", "sub_iv", "group_by",
              "mean_var", "group_mean_var", "se", "subject",
              "Title", "Legend Title", "X-axis Label", "Y-axis Label",
              "taskname", c(-10, 10), 8, 6, "keyword", c("blue", "red"), 5, "output.png")
 
[Example](https://github.com/jungheejung/cueR/blob/main/man/figures/example_plot_binned_rating.png)
}
