% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_halfrainclouds_onefactor.R
\name{plot_halfrainclouds_onefactor}
\alias{plot_halfrainclouds_onefactor}
\title{Plot Half Rainclouds for One Factor}
\usage{
plot_halfrainclouds_onefactor(
  subjectwise,
  groupwise,
  iv,
  sub_mean,
  group_mean,
  se,
  subject,
  ggtitle,
  title,
  xlab,
  ylab,
  taskname,
  ylim,
  w,
  h,
  dv_keyword,
  color,
  plot_savefname
)
}
\arguments{
\item{subjectwise}{A data frame containing subject-wise data.}

\item{groupwise}{A data frame containing group-wise summary data.}

\item{iv}{Independent variable name (as a string) in the data frame.}

\item{sub_mean}{The name of the subject-wise mean variable in the data frame.}

\item{group_mean}{The name of the group mean variable in the data frame.}

\item{se}{Standard error for the group mean.}

\item{subject}{The name of the subject identifier variable in the data frame.}

\item{ggtitle}{Title for the ggplot.}

\item{title}{Title for the legend.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{taskname}{Name of the task (unused in function).}

\item{ylim}{y-axis limits for the plot.}

\item{w}{Width of the}

\item{h}{Height of the saved plot.}

\item{dv_keyword}{Keyword associated with the dependent variable (unused in function).}

\item{color}{Color scheme for the plot, specified as a vector of colors.}

\item{plot_savefname}{Filename for saving the plot.}
}
\description{
This function creates a half raincloud plot for a given one-factor dataset. It combines violin plots,
box plots, and individual data points to provide a comprehensive view of the data distribution.
}
