% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_largetext.R
\name{plot_largetext}
\alias{plot_largetext}
\title{Apply a large text theme to a ggplot object}
\usage{
plot_largetext(ggplot_obj, text_size = 15, subtitle_size = 11)
}
\arguments{
\item{ggplot_obj}{A ggplot object to which the theme should be applied.}

\item{text_size}{Size of the main text elements in the plot; default is 15.}

\item{subtitle_size}{Size of the subtitle text elements in the plot; default is 11.}
}
\value{
A ggplot object with the updated theme.
}
\description{
This function takes a ggplot object and applies a theme that sets the text size and the subtitle size
with default values that can be customized.
}
\examples{
# Apply the custom theme with default sizes
my_ggplot <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
my_ggplot <- plot_largetext(my_ggplot)

# Apply the custom theme with specified text and subtitle sizes
my_ggplot <- plot_largetext(my_ggplot, text_size = 12, subtitle_size = 9)
}
