% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lineplot_onefactor.R
\name{plot_lineplot_onefactor}
\alias{plot_lineplot_onefactor}
\title{Plot a Line Plot for One Factor}
\usage{
plot_lineplot_onefactor(df, iv, mean, error, color, xlab, ylab, ggtitle)
}
\arguments{
\item{iv}{A string specifying the name of the independent variable column in `data`.}

\item{mean}{A string specifying the name of the column in `data` that contains mean values.}

\item{error}{A string specifying the name of the column in `data` that contains error margins.}

\item{color}{A vector specifying the color(s) to use for the plot lines.}

\item{xlab}{A string for the x-axis label.}

\item{ylab}{A string for the y-axis label.}

\item{ggtitle}{A string for the plot title.}

\item{data}{A data frame containing the data to plot.}
}
\value{
A ggplot object representing the line plot.
}
\description{
This function creates a line plot with error bars for a specified factor from the given dataset.
It is designed to plot mean values with associated error margins (like confidence intervals) for a single task.
}
\examples{
# Example usage:
# plot_lineplot_onefactor(my_data, "Task 1", "Group", "MeanScore", "ErrorMargin", "blue", "X-Axis", "Y-Axis", "Plot Title")

}
