% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lineplot_twofactor_subset.R
\name{plot_lineplot_twofactor_subset}
\alias{plot_lineplot_twofactor_subset}
\title{Plot a Two-Factor Line Plot with Optional Subsetting}
\usage{
plot_lineplot_twofactor_subset(
  data,
  taskname,
  iv1,
  iv2,
  mean,
  error,
  color,
  ggtitle,
  xlab = "Stimulus intensity",
  ylab = "Outcome rating",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{data}{A data frame containing the dataset to be plotted.}

\item{taskname}{A string specifying the task name for subsetting the data, or NULL to use the entire dataset.}

\item{iv1}{The name of the first independent variable (factor) for the x-axis.}

\item{iv2}{The name of the second independent variable (factor) used for grouping and coloring.}

\item{mean}{The name of the column containing mean values to be plotted on the y-axis.}

\item{error}{The name of the column containing error values for error bars.}

\item{color}{A vector of colors to be used for the different groups represented by `iv2`.}

\item{ggtitle}{The title of the plot.}

\item{xlab}{Label for the x-axis. Default is "Stimulus Intensity".}

\item{ylab}{Label for the y-axis. Default is "Outcome Rating".}
}
\value{
A ggplot object representing the line plot.
}
\description{
This function creates a line plot for data with two factors. It allows for optional subsetting of the data based on a specified task name. If the 'taskname' column does not exist or 'taskname' is NULL, the function uses the entire dataset. It utilizes ggplot2 for plotting.
}
\examples{
# Example usage with taskname specified:
p1 = plot_lineplot_twofactor_subset(data = mydata, taskname = "task1",
              iv1 = "intensity", iv2 = "cue", mean = "mean_score", error = "ci",
              color = c("red", "blue"), ggtitle = "My Plot")

# Example usage without taskname (using entire dataset):
p2 = plot_lineplot_twofactor_subset(data = mydata, taskname = NULL,
              iv1 = "intensity", iv2 = "cue", mean = "mean_score", error = "ci",
              color = c("red", "blue"), ggtitle = "My Plot")
[Example plot](TODO)
}
