% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signature_twolineplot.R
\name{plot_signature_twolineplot}
\alias{plot_signature_twolineplot}
\title{Plot Signature with Two-Line Plot}
\usage{
plot_signature_twolineplot(signature_key = NULL, plot_keys = NULL, df = NULL)
}
\arguments{
\item{signature_key}{A keyword associated with the signature to be plotted.}

\item{plot_keys}{A list containing keys for various parameters required for plotting and computation, such as 'iv1', 'iv2', 'sub_mean', 'sub', 'dv', 'group_mean', 'error', 'color', 'ggtitle', 'xlab', and 'ylab'.}

\item{df}{The data frame containing the data to be used for plotting.}
}
\value{
A ggplot object representing the two-line plot based on the specified signature.
}
\description{
This function creates a two-line plot for a specific signature based on provided keys and a dataframe. It sources additional R scripts, computes a summary with two factors, and then uses these summaries to generate the plot.
}
\examples{
# Assuming 'plot_keys' and 'df' are predefined and appropriate
plot_signature_twolineplot(signature_key = "NPSpos", plot_keys = plot_keys, df = df)
}
