% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signature_tworaincloud.R
\name{plot_signature_tworaincloud}
\alias{plot_signature_tworaincloud}
\title{Plot Signature with Two-Factor Raincloud Plots}
\usage{
plot_signature_tworaincloud(signature_key = NULL, plot_keys = NULL, df = NULL)
}
\arguments{
\item{signature_key}{(Optional) A key or identifier for the signature being plotted.}

\item{plot_keys}{A list of keys specifying the plotting parameters and variables to be used.
This includes variables for mean per subject, subject identification, independent
variables, dependent variable, plot titles, labels, and other plotting parameters.

Important keys include:
- `sub_mean`: Mean value per subject.
- `iv1`: First independent variable.
- `iv2`: Second independent variable.
- `dv`: Dependent variable.
- `ggtitle`: Title for the plot.
(additional keys can be found in the documentation for `plot_halfrainclouds_twofactor`)}

\item{df}{A dataframe containing the data to be plotted.}
}
\value{
A ggplot object representing the two-factor raincloud plot.
}
\description{
This function creates raincloud plots for a specified signature key with data
grouped by two factors. It first computes summary statistics and then uses these
to generate the plots.
}
\examples{
# Assuming appropriate data and plot_keys
# p <- plot_signature_tworaincloud(signature_key = "signature1",
#                                  plot_keys = list_of_keys, df = mydata)

}
