% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_bar_grayarrange.R
\name{plot_timeseries_bar_grayarrange}
\alias{plot_timeseries_bar_grayarrange}
\title{Plot Time Series Bar with Gray Arrange}
\usage{
plot_timeseries_bar_grayarrange(
  df,
  iv1,
  iv2,
  mean,
  error,
  xlab,
  ylab,
  ggtitle,
  color
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{iv1}{The name of the first independent variable for the x-axis.}

\item{iv2}{The name of the second independent variable used for grouping and coloring, with special handling for 'gray' levels.}

\item{mean}{The name of the column containing mean values to be plotted on the y-axis.}

\item{error}{The name of the column containing error values for error bars.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{ggtitle}{Title of the plot.}

\item{color}{A named vector of colors to be used for the different groups represented by `iv2`.}
}
\value{
A ggplot object representing the time series bar plot.
}
\description{
This function creates a time series bar plot with a specific arrangement for gray and non-gray levels of a factor. It uses ggplot2 for plotting and allows for custom color settings. The levels of the second independent variable are reordered so that gray levels appear first.
}
\examples{
df <- data.frame(
    time = c(1, 2, 3, 1, 2, 3),
    condition = c("gray1", "gray2", "non-gray1", "gray1", "gray2", "non-gray1"),
    mean = rnorm(6),
    error = runif(6, 0.1, 0.5)
)
plot_timeseries_bar_grayarrange(df, "time", "condition", "mean", "error",
    "Time", "Response", "My Plot", c("gray1" = "gray", "gray2" = "darkgray", "non-gray1" = "blue"))
}
