% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_printdash.R, R/print_dash.R
\name{print_dash}
\alias{print_dash}
\title{Print a title surrounded by dashes for emphasis.}
\usage{
print_dash(title)

print_dash(title)
}
\arguments{
\item{title}{The title to be printed.}
}
\value{
This function does not return a value; it prints the title with dashes.

This function does not return a value; it prints the title with dashes.
}
\description{
This function prints a title surrounded by dashes to create a visual separation
and emphasis for the printed output.

This function prints a title surrounded by dashes to create a visual separation
and emphasis for the printed output.
}
\examples{
\dontrun{
  print_dash("model with Z scores")
}

\dontrun{
  print_dash("model with Z scores")
}

}
