% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_contrast_beh.R
\name{simple_contrast_beh}
\alias{simple_contrast_beh}
\title{Add Contrast and Ordered Categorical Columns to a Data Frame}
\usage{
simple_contrast_beh(df)
}
\arguments{
\item{df}{A data frame to which new columns will be added. The data frame is expected to have
specific columns like `src_subject_id`, `param_stimulus_type`, and `param_cue_type`.}
}
\value{
A modified data frame with added columns for contrasts and ordered categoricals.
        The function also prints out a message listing the names of the new columns.
}
\description{
This function takes a data frame and adds several new columns to it. These new columns include
contrast codes (linear and quadratic), ordered categorical variables, and factorized variables.
The function is specifically tailored for a behavioral file with a fixed structure.
}
\examples{
# Assuming `data` is your data frame with the required structure:
modified_data <- simple_contrast_beh(data)

}
