% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_contrast_singletrial.R
\name{simple_contrast_singletrial}
\alias{simple_contrast_singletrial}
\title{Apply Simple Contrasts to Single Trial Data}
\usage{
simple_contrast_singletrial(df)
}
\arguments{
\item{df}{A data frame containing the single trial data. 
It must have columns 'stimintensity' and 'cuetype'.}
}
\value{
The modified data frame with added columns for contrasts:
        - `STIM` as a factor based on 'stim'.
        - `STIM_linear` for linear contrast coding based on 'stimintensity'.
        - `STIM_quadratic` for quadratic contrast coding based on 'stimintensity'.
        - `CUE_high_gt_low` for cue contrast between high and low cue types.
        - `stim_ordered` as an ordered factor for 'stimintensity'.
        - `cue_name` and `cue_ordered` as factors for 'cuetype'.
}
\description{
This function applies simple linear and quadratic contrasts to single trial data based on stimulus intensity and cue type. It modifies the input dataframe by adding several new columns representing these contrasts.
}
\examples{
df <- data.frame(
    stim = c("A", "B", "C"),
    stimintensity = c("low", "med", "high"),
    cuetype = c("cuetype-low", "cuetype-med", "cuetype-high")
)
df <- simple_contrast_singletrial(df)

}
