% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Summary Statistics with Standard Error and Confidence Interval}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{A data frame containing the data to be summarized.}

\item{measurevar}{The name of the column in `data` that contains the variable to be summarized.}

\item{groupvars}{A vector containing the names of columns in `data` that contain grouping variables.
If NULL, the function will summarize the entire dataset without grouping.}

\item{na.rm}{Logical. Indicates whether to ignore NA values in `measurevar`. 
Defaults to FALSE.}

\item{conf.interval}{The confidence level for the interval. The default is 0.95 
for a 95% confidence interval.}

\item{.drop}{Logical. If TRUE (default), will drop levels that do not occur in 
the data frame `groupvars`.}
}
\value{
A data frame with the following columns: number of observations (N), 
  mean, median, standard deviation (sd), standard error (se), and confidence interval (ci).
}
\description{
Calculates and returns summary statistics for a given variable in a data frame, 
grouped by one or more variables. It provides count, mean, median, standard deviation, 
standard error of the mean, and confidence interval (default 95%).
}
\examples{
# Example usage with mtcars dataset
summarySE(mtcars, measurevar="mpg", groupvars="cyl")

}
\references{
This code is aa direct copy of the function used in RaincloudPlots: 
https://github.com/RainCloudPlots/RainCloudPlots/blob/master/tutorial_R/summarySE.R
This version is an Adapted code of Ryan Hope's Rmis::summarySE function: 
https://www.rdocumentation.org/packages/Rmisc/versions/1.5/topics/summarySE
}
