% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySEwithin.R
\name{summarySEwithin}
\alias{summarySEwithin}
\title{Summarizes Data with Within-Subjects Variables}
\usage{
summarySEwithin(
  data = NULL,
  measurevar,
  betweenvars = NULL,
  withinvars = NULL,
  idvar = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{A data frame containing the data to be summarized.}

\item{measurevar}{The name of a column that contains the variable to be summarized.}

\item{betweenvars}{A vector containing names of columns that are between-subjects variables.}

\item{withinvars}{A vector containing names of columns that are within-subjects variables.}

\item{idvar}{The name of a column that identifies each subject (or matched subjects).}

\item{na.rm}{A boolean indicating whether to ignore NA's in calculations.}

\item{conf.interval}{The percent range of the confidence interval (default is 95%).}

\item{.drop}{A boolean indicating whether to drop levels that do not appear in the data.}
}
\value{
A data frame with summarized statistics including mean, standard deviation,
        standard error, and confidence interval for both normed and un-normed data.
}
\description{
This function summarizes data, especially handling within-subjects variables by removing
inter-subject variability. It calculates count, un-normed mean, normed mean, standard
deviation, standard error of the mean, and confidence interval. Adjusted values are calculated
using the method from Morey (2008) if there are within-subject variables.
}
\examples{
# Assuming `data` is your data frame with the appropriate structure:
result <- summarySEwithin(data, "measurevar", c("betweenVar1", "betweenVar2"),
                          c("withinVar1", "withinVar2"), "idvar", TRUE, .95, TRUE)
}
\references{
This code is aa direct copy of the function used in RaincloudPlots:
https://github.com/RainCloudPlots/RainCloudPlots/blob/master/tutorial_R/summarySE.R
This version is an Adapted code of Ryan Hope's Rmis::summarySE function:
https://www.rdocumentation.org/packages/Rmisc/versions/1.5/topics/summarySE
}
